Refactor package structure to separate constructor based bean injection code set from setter based bean injection code set
