+++ spring-core/src/main/java/com/baeldung/beaninjection/constructorbased/ChatMessage.java
@@ -1,4 +1,4 @@
-package com.baeldung.beaninjection;
+package com.baeldung.beaninjection.constructorbased;
 
 public interface ChatMessage {
     public void send();
