+++ spring-core/src/main/java/com/baeldung/beaninjection/constructorbased/MessageSender.java
@@ -1,4 +1,4 @@
-package com.baeldung.beaninjection;
+package com.baeldung.beaninjection.constructorbased;
 
 public class MessageSender {
